Revert "Adding Demo Code Schema"
This reverts commit 309d203c877abc83b81df24c1fc5a5af9b3567fb.
