+++ Client/Pages/Repl.razor.cs
@@ -33,6 +33,8 @@
         [Parameter]
         public int? DemoId { get; set; }
 
+        private DotNetObjectReference<Repl> DotNetInstance { get; set; }
+
         public string DemoCode { get; set; }
 
         public CodeEditor CodeEditor { get; set; }
@@ -50,8 +52,6 @@
         public bool Loading { get; set; }
 
         public int UserComponentCodeStartLine => this.UserComponentCodePrefix.Count(ch => ch == '\n');
-
-        private DotNetObjectReference<Repl> DotNetInstance { get; set; }
 
         public async Task UpdateLoaderText(string loaderText)
         {
